+++ Unit Armies/ES.h
@@ -6,7 +6,7 @@ class ES :public Unit {
 
 
 public:
-	ES(int id = 0, int tj = 0, double health = 0, float power = 0, int attackCap = 0,
+	ES(int id = 0, int tj = 0, int health = 0, int power = 0, int attackCap = 0,
 		std::string type = "");
 
 	void attack();
