+++ Unit Armies/ET.cpp
@@ -1,6 +1,6 @@
 #include "ET.h"
 
-ET::ET(int id, int tj, double health, float power, int attackCap,
+ET::ET(int id, int tj, int health, int power, int attackCap,
 	std::string type) :Unit(id, tj, health, power, attackCap, type)
 {
